lab1/Lab1/src/main/java/package1/Author.java
package package1;

public class Author {
    private String name;
    private String email;
    private String address;

        public Author (String name, String email, String address){
            this.name =name;
            this.email =email;
            this.address =address;
        }

        public String getName(){
            return name;
        }

        public String getEmail(){
            return email;
        }

        public String getAddress(){
            return address;
        }

}
